Bug 1708692 - Rearrange arena header to move alloc kind out of flags bitfield r=sfink
This patch moves the alloc kind out of the flags word and places it between the
freespan and the zone. On 64 bit platforms there is 32 bits of padding here
because zone has pointer alignment and the freespan is 32 bits. On 32 bits
platforms this increases the size of the header but there is enough spare that
this doesn't reduce the arena density for any of our alloc kinds.

This also lets us type allocKind appropraitely.

Depends on D114213

Differential Revision: https://phabricator.services.mozilla.com/D114214

+++ js/src/gc/Heap-inl.h
@@ -31,7 +31,7 @@ inline void js::gc::Arena::init(JS::Zone* zoneArg, AllocKind kind,
   MOZ_MAKE_MEM_UNDEFINED(this, ArenaSize);
 
   zone = zoneArg;
-  allocKind = size_t(kind);
+  allocKind = kind;
   onDelayedMarkingList_ = 0;
   hasDelayedBlackMarking_ = 0;
   hasDelayedGrayMarking_ = 0;
